refactor: split SessionsTab into sub-components
